requirements.txt
numpy==1.16.0
tensorflow-gpu==1.8
progressbar2==3.37.1
librosa==0.6.0
ffmpeg==1.4
pyworld==0.2.8
wget==3.2
tqdm==4.31.1